FoolHD/requirements.txt
adversarial-robustness-toolbox==1.4.1
kaldiio==2.14.1
numpy==1.16.4
pandas==1.1.2
sklearn==0.0
scipy==1.4.1
torch-stft==0.1.4
torchaudio==0.4.0
tqdm==4.19.8}
